+++ app/helpers/application_helper.rb
@@ -101,17 +101,16 @@ module ApplicationHelper
       text = "<p>This manual has a duplicate slug and can't be published.</p>"
     else
       text = ""
-      update_type = ManualUpdateType.for(manual)
-      if update_type == "minor"
+      if manual.update_type == "minor"
         text += "<p>You are about to publish a <strong>minor edit</strong>.</p>"
-      elsif update_type == "major" && manual.has_ever_been_published?
+      elsif manual.update_type == "major" && manual.has_ever_been_published?
         text += "<p><strong>You are about to publish a major edit with public change notes.</strong></p>"
       end
       if manual.use_originally_published_at_for_public_timestamp? && manual.originally_published_at.present?
         text += "<p>The updated timestamp on GOV.UK will be set to the first publication date.</p>"
-      elsif update_type == "minor"
+      elsif manual.update_type == "minor"
         text += "<p>The updated timestamp on GOV.UK will not change.</p>"
-      elsif update_type == "major"
+      elsif manual.update_type == "major"
         text += "<p>The updated timestamp on GOV.UK will be set to the time you press the publish button.</p>"
       end
     end
